Allow admin root URL to be at '/' instead of '/admin/'
The previous implementation of `get_native_model_name()` relied on the fact that the first two `url_parts` were assumed to be `['', 'admin']`. But if one configures the admin to be located at `/` instead of the default `/admin/`, that is not true and no valid model will be returned. The same should be true for setups where the admin is at something like `/custom/admin/`.