Update form_tabs.rst
Edited Lines 37,41,44 so each tab is a single string object rather than a whole. When whole and no `suit_form_includes` is given (probably) the tabs wont hide the first tab (e.g. when in cities, the contents of general will be on the bottom as an inline form and not hidden) The edit solves this issue.