Fixed issue where SortableGenericInlines would break whenever adding blank inline forms.
This is because inline fieldset names append the names of the generic relation fields as defined in models.py. For instance, instead of being called "model_set-n-field" they'd be called
"app-model-content_type-object_id-n-field" reflecting the names of the generic type and id fields. I replaced the string split on '-' with a regexp split that looks for the -n- pattern instead.

This appears to work in Chrome for both regular and generic relations.