Better handling of checkboxes in sortable inline forms.
Previously sortable.js was unable to detect whether or not a checkbox had deviated from its original value. First of all the test to find non-blank values:

$set_block.find(":input[value!=''][type!='hidden']")

would not work on a checkbox, since its value will evaluate to "on" or "off".
Secondly, the function filter_unchanged only handled select boxes and the sortable field. As a result, a checkbox in an inline form would always flag a row as "changed". This is particularily troublesome when you have extra blank forms in a formset.

The fix was to add a block in filter_unchanged to check for a checkbox field, and compare its current checked property to that of defaultChecked.