Fix for jQuery 3 removal of .selector tag
jQuery 3 removed the .selector tag, which means that for any modern version of Django the sortable code does not work. Rather than relying on the .selector, this patch follows the recommended jQuery practice of allowing the selector to be passed as an argument. As the suit_list_sortable is not called anywhere except from this function, this fix should be all that is needed.